requirements.txt
numpy==1.24.3
scipy==1.11.4
pandas==2.0.3

tqdm==4.66.5

# tensorflow==2.13.1

scikit-learn==1.3.2
scikit-image==0.22.0

imageio==2.35.1
Pillow==10.4.0

medpy==0.5.2
SimpleITK==2.3.1
torchinfo  # <--- 요청하신 torchinfo 추가
tensorboard
matplotlib==3.8.4